pyproject.toml
[project]
name = "saqr-mcp"
version = "0.1.0"
description = "Saqr-MCP is a Python application that implements the Model Context Protocol (MCP) to enable AI assistant capabilities with local models"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq==0.24.0",
    "httpx==0.28.1",
    "loguru==0.7.3",
    "mcp[cli]==1.7.1",
    "ollama==0.4.8",
    "tavily-python==0.7.2",
]
